+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.1.1
+## Changes
+1. Replace SoftMaker Office to LibreOffice
+2. Add Firefox, Thuderbird and Telegram that on disk K
+
 # v0.1
 ## Changes
 - First release